requirements.txt

Django==4.2
python-decouple==3.8
djangorestframework==3.15.2
drf-yasg==1.21.7
django-cors-headers==4.4.0
django-storages==1.14.4
psycopg2-binary==2.9.9
requests==2.32.3
djangorestframework-simplejwt==5.5.0
